chore: Moved one extra writer into the graph module
