+++ ra2ce/common/README.md
@@ -1,0 +1,3 @@
+# Common
+
+This module is meant to contain **ONLY** protocols (`typing.Protocol`) and very concrete readers / writers that could be used both by an analysis ( `race.analyses`) or a graph ( `race.graph`).
