+++ ra2ce/graph/networks.py
@@ -37,7 +37,7 @@ from ra2ce.graph.network_config_data.network_config_data import NetworkConfigDat
 from ra2ce.graph import networks_utils as nut
 from ra2ce.graph.segmentation import Segmentation
 from ra2ce.io.readers import GraphPickleReader
-from ra2ce.io.writers import JsonExporter
+from ra2ce.graph.exporters.json_exporter import JsonExporter
 from ra2ce.graph.exporters.network_exporter_factory import NetworkExporterFactory
